Moved bidirectional_map.hpp to exchcxx/util
